+++ src/sox_private.h
@@ -2,17 +2,14 @@
 #ifndef _SOX_PRIVATE_H_
 #define _SOX_PRIVATE_H_
 
-// input 
-extern float *g_cur;
-extern uint16_t *g_celVol;
-extern int16_t *g_celTmp;
-// output
-extern uint16_t *g_celSOC;
-extern uint16_t *g_grpSOC;
-extern float *g_celSOH;
-extern float *g_grpSOH;
-
-extern uint16_t *g_cycleCount;
+extern float *g_cur;                       // A            *1
+extern uint16_t *g_celVol;                 // mv           *1
+extern int16_t *g_celTmp;                  // dregree      *10
+extern uint16_t *g_celSOC;                 // %            *10
+extern uint16_t *g_grpSOC;                 // %            *10
+extern double  *g_celSOH;                  // %            *1
+extern double  *g_grpSOH;                  // %            *1
+extern uint32_t *g_cycleCount;             // times        *1000
 
 
 #endif
